Move reducer tests.
